[clear]state管理のファイルを一元化
プロジェクトも小さいのでstate管理のファイルを一元化
