--- front/lib/atoms/composingState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const composingState = atom({
-  key: 'composing',
-  default: false,
-});
