--- front/lib/atoms/currentCityState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const currentCityStat = atom({
-  key: 'currentCityStat',
-  default: '',
-});
