--- front/lib/atoms/dayWeatherState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const dayWeatherState = atom({
-  key: 'dayWeatherState',
-  default: [],
-});
