--- front/lib/atoms/disabledState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const disabledState = atom({
-  key: 'disabledState',
-  default: true,
-});
