--- front/lib/atoms/errorMessageState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const errorMessageState = atom({
-  key: 'errorMessageState',
-  default: '',
-});
