--- front/lib/atoms/getErrorState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const getErrorState = atom({
-  key: 'getError',
-  default: false,
-});
