--- front/lib/atoms/inputCityState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const inputCityState = atom({
-  key: 'inputCityState',
-  default: '',
-});
