--- front/lib/atoms/isLoadingState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const isLoadingState = atom({
-  key: 'isLoading',
-  default: false,
-});
