--- front/lib/atoms/timeWeatherState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const timeWeatherState = atom({
-  key: 'timeWeatherState',
-  default: [],
-});
